Fix pooled_hazard_task bug
Pooled hazards task does not work as intended (or errors) when using tasks that have non-null row_index internal variables. This occurs because when passing in a new dataset to next_in_chain, it does not reset the internal row_index variable. This makes using CV learners that utilize pooled hazards tasks internally (e.g. Lrnr_cv$new(Lrnr_pooled_hazards$new())) break down. For tmle3 survival, the pooled hazards task was created externally as the main task, so the hazard estimation with CV was not impacted by this bug. 

It might also be worth changing sl3_Task so that row_index is reset when passing in a new dataset. Or, maybe we should just not allow datasets to be passed in through next_in_chain.